+++ homework-08/src/main/java/ru/otus/homework/controller/CommentsCommandsController.java
@@ -2,6 +2,7 @@ package ru.otus.homework.controller;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.shell.standard.ShellComponent;
@@ -27,6 +28,9 @@ public class CommentsCommandsController {
 
   @ShellMethod(value = "Add comment. Enter bookId, text", key = {"addComment", "ac"})
   public String add(String bookId, String text) {
+    if (checkIdForInvalidity(bookId)) {
+      return "Invalid id";
+    }
     CommentDto insertedComment;
     try {
       insertedComment = commentService.add(bookId, text);
@@ -38,6 +42,9 @@ public class CommentsCommandsController {
 
   @ShellMethod(value = "Get comment. Enter id", key = {"getComment", "gc"})
   public String get(String id) {
+    if (checkIdForInvalidity(id)) {
+      return "Invalid id";
+    }
     CommentDto comment;
     try {
       comment = commentService.get(id);
@@ -49,6 +56,9 @@ public class CommentsCommandsController {
 
   @ShellMethod(value = "Get comments by book id", key = {"getCommentsByBookId", "gcb"})
   public String getByBookId(String id) {
+    if (checkIdForInvalidity(id)) {
+      return "Invalid id";
+    }
     List<CommentDto> comments = commentService.getByBookId(id);
     return comments.isEmpty() ? "There is no comments for this book present" :
         comments.stream()
@@ -58,6 +68,9 @@ public class CommentsCommandsController {
 
   @ShellMethod(value = "Update comment. Enter id, bookId, text", key = {"updateComment", "uc"})
   public String update(String id, String bookId, String text) {
+    if (checkIdForInvalidity(id, bookId)) {
+      return "Invalid id";
+    }
     CommentDto commentDto;
     try {
       commentDto = commentService.update(id, bookId, text);
@@ -69,7 +82,19 @@ public class CommentsCommandsController {
 
   @ShellMethod(value = "Remove comment. Enter id", key = {"removeComment", "rc"})
   public String remove(String id) {
+    if (checkIdForInvalidity(id)) {
+      return "Invalid id";
+    }
     commentService.remove(id);
     return "Comment with id " + id + " removed";
+  }
+
+  private boolean checkIdForInvalidity(String ... ids) {
+    for (String id : ids) {
+      if (!ObjectId.isValid(id)) {
+        return true;
+      }
+    }
+    return false;
   }
 }
